Final NerdSwag Front-End E-commerce shop project
I fixed the cart component to make purchases work properly and to erase products after a successful purchase. Also, I properly implemented the different resulting responses to the user's attempted purchases. I made the cart component show the remaining balance for a purchase before attempting the purchase in the cart. The remaining balance displays properly after a purchase. I implemented a limit to options depending if the user is logged in or not. If the user is not logged in, then the wishlist, cart, and add to cart and wishlist buttons don't display. If the user is logged in, then they can use the add to cart and add to wishlist buttons and the wishlist as well as the cart components display on screen. I added the feature of editing the balance of the user's account. I also added the features of the cart and wishlist components displaying what products are in them based on what the user left in their wishlist and cart respectively through their account. Lastly, I replaced the alerts with a pop-up module that notifies the user on the result of their actions for stylization purposes.